src/main/java/Seminar1/calculator/Calculator.java
package Seminar1.calculator;
public class Calculator {
    /**
     Задание 1. ** В классе Calculator создайте метод calculateDiscount, который принимает сумму покупки и процент скидки и возвращает
     сумму с учетом скидки. Ваша задача - проверить этот метод с использованием библиотеки AssertJ. Если метод calculateDiscount получает
     недопустимые аргументы, он должен выбрасывать исключение ArithmeticException. Не забудьте написать тесты для проверки этого поведения.
     */

    public double calculateDiscount(double summ, int percent) {
        if (summ < 0 || percent < 0) {
            throw new IllegalArgumentException("the purchase amount or discount should not be negative");
        }
        return summ * percent/100;
    }

    public int add(int a, int b) {
        return a + b;
    }

    public int subtract(int a, int b) {
        return a - b;
    }

    public int multiply(int a, int b) {
        return a * b;
    }

    public double divide(int a, int b) {
        if (b == 0) {
            throw new IllegalArgumentException("Divider cannot be zero");
        }
        return (double) a / b;
    }

}

